Replace `brew tap` call with directory check (#5)
`brew tap` is really slow (#4), which affects loading time in default
Homebrew environments.

A workaround is to declare `CHTF_AUTO_INSTALL_METHOD` before loading,
but let's avoid the invocation in all cases. The file hierarchy for Taps
can be assumed to be stable enough to check the existence by ourselves.

We still call `brew --caskroom` and `brew --repo`, but they are fast.